chore(rebase): resolve conflicts from project restructure
